Fix: Compiler warning C4244 'argument': conversion from 'WPARAM' to 'const int32_t', possible loss of data

+++ SDF/src/Core/Window.cpp
@@ -168,7 +168,7 @@ namespace sdf {
             (isPressed << 3),
             GET_X_LPARAM(lParam),
             GET_Y_LPARAM(lParam),
-            Input::toButton(wParam)
+            Input::toButton((int32_t)wParam)
         ));
 
     }
